doing some tests
trying to get this to work currently getting a 500 error, unclear why.

+++ DuggaSys/tests/microservices/adduser_ms_test.php
@@ -4,15 +4,15 @@ include "../../../Shared/test.php";
 
 $testsData = array(
     //------------------------------------------------------------------------------------------
-    // This tests the microservice updateActiveUsers_ms 
+    // This tests the microservice addUser_ms 
     //------------------------------------------------------------------------------------------
-    'updateActiveUsers_ms' => array(
+    'addUser_ms' => array(
         'expected-output' => '9999',
         'query-before-test-1' => "INSERT INTO course (cid,coursecode,coursename,visibility,creator, hp, courseGitURL) VALUES(999,'test','toBeDeleted', 0, 1, 7.5, null)",
         'query-before-test-2' => "INSERT INTO user (uid, username, email, firstname, lastname, ssn, password,addedtime, class) VALUES(9999 , 'hacker', 'hacker@student.his.se', 'tester', 'test', 00000000-0000, 'test',now(), WEBUG14h);",
         'query-after-test-1' => "DELETE FROM user WHERE uid = 9999;",
         'query-after-test-2' => "DELETE FROM course WHERE cid = 999;",
-        'query-after-test-3' => "DELETE FROM user_course WHERE cid = 999;",
+        'query-after-test-3' => "DELETE FROM user_course WHERE cid = 999 AND uid = 9999;"
         'service' => 'http://localhost/LenaSYS/DuggaSys/microservices/accessedService/addUser_ms.php',
         'service-data' => serialize(
             array( // Data that service needs to execute function
@@ -22,8 +22,8 @@ $testsData = array(
                 'courseid' => '999',
                 'newusers' =Z 'hacker@student.his.se',
                 'coursevers' => '97732',
-                'hash' => 'hjk4ert6',
-                'AUtoken' => 999
+                'hash' => 'hjk4ert6'
+                // 'AUtoken' => 999
             )
         ),
         'filter-output' => serialize(
